fix(booking): adjust booking price based on the vehicle

+++ src/main/java/com/ul/vrs/entity/booking/decorator/InsuranceBookingDecorator.java
@@ -8,7 +8,7 @@ public class InsuranceBookingDecorator extends BookingDecorator{
     }
 
     // TODO: Adjust value based on real-life values
-    public long getPrice() {
+    public double getPrice() {
         return super.getPrice() + 100;
     }
 }
